+++ test-studio/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Fix: account for new Sanity version that broke useMemoObservable, used in document previews
 - Updated dependencies
-  - sanity-plugin-cloudflare-r2-files@0.1.1
+  - sanity-plugin-r2-files@0.1.0
   - sanity-plugin-digital-ocean-files@1.0.2
   - sanity-plugin-external-files@1.0.2
   - sanity-plugin-s3-files@1.0.2